requirements.txt
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-crontab==0.7.1
django-phonenumber-field==7.1.0
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.19
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
urllib3==2.0.4
wcwidth==0.2.6
